+++ Coinbase.AdvancedTradeApiClient/Models/Internal/InternalCandle.cs
@@ -55,10 +55,10 @@ internal class InternalCandle : IModelMapper<Candle>
         return new Candle
         {
             StartUnix = StartUnix,
-            Low = Low.ToDouble(),
-            High = High.ToDouble(),
-            Open = Open.ToDouble(),
-            Close = Close.ToDouble(),
+            Low = Low.ToDecimal(),
+            High = High.ToDecimal(),
+            Open = Open.ToDecimal(),
+            Close = Close.ToDecimal(),
             Volume = Volume.ToDouble(),
         };
     }
